refactor: move RegistryDriverIcon out of the connection-modal folder
RegistryDriverIcon (a generic <img>-with-fallback for a registry-hosted
icon URL) lived under src/components/modals/connection/, used only by that
folder's two components (EngineCard, InstallGate). Moves it to
src/components/ — matching ConnectionIconImage's sibling placement — so
src/utils/driverUI.tsx (a general utility, not modal-specific) can reuse it
in the next commit without reaching into a single modal's private folder.

No behavior change; just updates the two existing importers' relative paths.
