+++ src/components/RegistryDriverIcon.tsx
